SHA-1930 : Unable to delete folder, when the folder contains shortcut link to the same folder    - Check link for PENDING_DELETE aspect before trying to delete it    - Added unit test for case
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/BRANCHES/DEV/5.2.N/root@133020 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/doclink/DocumentLinkServiceImpl.java
@@ -305,10 +305,13 @@ public class DocumentLinkServiceImpl implements DocumentLinkService, NodeService
         {
             try
             {
-                nodeService.deleteNode(linkRef);
+                if (!nodeService.hasAspect(linkRef, ContentModel.ASPECT_PENDING_DELETE))
+                {
+                    nodeService.deleteNode(linkRef);
 
-                /* if the node was successfully deleted increment the count */
-                report.incrementDeletedLinksCount();
+                    /* if the node was successfully deleted increment the count */
+                    report.incrementDeletedLinksCount();
+                }
             }
             catch (AccessDeniedException ex)
             {
